requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.10.0
asn1crypto==1.5.1
asttokens==3.0.1
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
comm==0.2.3
cryptography==46.0.3
cssselect2==0.8.0
debugpy==1.8.19
decorator==5.2.1
Django==5.2.8
executing==2.2.1
freetype-py==2.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-speech==2.34.0
google-cloud-texttospeech==2.33.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
html5lib==1.1
httplib2==0.31.0
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
lxml==6.0.2
matplotlib-inline==0.2.1
mysqlclient==2.2.7
nest-asyncio==1.6.0
oscrypto==1.3.0
packaging==25.0
parso==0.8.5
pillow==12.0.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.2.1
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyparsing==3.2.5
pypdf==6.5.0
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
reportlab==4.4.7
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
svglib==1.6.0
tinycss2==1.5.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.5.0
wcwidth==0.2.14
webencodings==0.5.1
xhtml2pdf==0.2.17
